% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_categoricos.R
\name{desc_categoricos}
\alias{desc_categoricos}
\title{Analisis descriptivo a variables categoricas}
\usage{
desc_categoricos(data)
}
\arguments{
\item{data}{DataFrame con las variables categoricas a visualizar.}
}
\value{
Graficos y resultados
}
\description{
Esta funcion realiza un grafico resumen (proporciones) a un dataframe con variables categoricas. En caso de ingresar variables continuas se llevaran a character y seran tratadas como categoricas.
}
\examples{
desc_categoricos(data = CO2[, 2:3])
desc_categoricos(data = mtcars[, c(2, 8:11)])
desc_categoricos(data = cars)
}
